Uptown/Trump Fixed

+++ core/src/com/zayacam/game/bidwhist/cards/Card.java
@@ -127,7 +127,7 @@ public class Card  {
                 this.cardFace = CardFace.Joker;
             	break;
             default:
-                result = Integer.toString(cardValue);
+                result = Float.toString(cardValue);
         }
         return result;
